refactor(backend): extract test result crud module
